+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Custom TailwindCSS/PostCSS entrypoint setting.
+
 ### Fixed
 
 - Removed 'obsidian' from plugin ID in `manifest.json`.